feat: moved models to data access and added base repo
