+++ core/data_access/base_repositories/base_repository.py
@@ -1,0 +1,27 @@
+# core/data_access/Base/base_repository.py
+from abc import ABC, abstractmethod
+from typing import Generic, TypeVar, List, Optional
+
+T = TypeVar('T')
+
+
+class BaseRepository(ABC, Generic[T]):
+    @abstractmethod
+    def get_all(self) -> List[T]:
+        pass
+
+    @abstractmethod
+    def get_by_id(self, id: int) -> Optional[T]:
+        pass
+
+    @abstractmethod
+    def create(self, obj: T) -> T:
+        pass
+
+    @abstractmethod
+    def update(self, obj: T) -> T:
+        pass
+
+    @abstractmethod
+    def delete(self, id: int) -> None:
+        pass
