+++ core/serializers/player_serializer.py
@@ -1,5 +1,5 @@
 from rest_framework import serializers
-from core.models.player_model import Player
+from core.data_access.models.player_model import Player
 from core.common.validators import validate_format, validate_alpha, validate_non_empty, validate_max_length, \
     validate_min_length
